+++ CHANGES.rst
@@ -4,7 +4,12 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- TBD
+- Add ``session_factory_from_settings`` and ``includeme``.
+  See https://github.com/Pylons/pyramid_nacl_session/pull/11
+
+- Drop Python 2.6, 3.3, and 3.4 support.
+
+- Add Python 3.6, 3.7 and 3.8 support.
 
 0.3 (2016-02-16)
 ----------------